renew dataset name as loader name| add trainer
