+++ requirements.txt
@@ -1,5 +1,5 @@
-torchvision>=0.5.0
-tensorflow>=2.6
+# torchvision>=0.5.0
+# tensorflow>=2.6
 einops
 
 xmltodict
